requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7
langchain-ollama>=0.3.7
langchain-huggingface>=0.3.1

# Vector stores and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Web search
tavily-python>=0.3.3

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Web interface
streamlit>=1.29.0

# Utilities
pydantic>=2.5.2
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0

# Evaluation
scikit-learn>=1.3.2